Create MANIFEST.in
Basic setup recommendations

A MANIFEST.in is needed when you need to package additional files that are not automatically included in a source distribution. To see a list of what’s included by default, see the Specifying the files to distribute section from the distutils documentation.

For an example, see the MANIFEST.in from the PyPA sample project.

For details on writing a MANIFEST.in file, see the The MANIFEST.in template section from the distutils documentation.